requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework
freezegun
numpy
pandas
black
pytest
pytest-django
pytest-cov
pytest-pythonpath
django-extensions
